Prompts user to try again if entered wrong command (#21)
* Prompts user to try again if entered wrong command

before it assumed the player was trying to run if entered wrong command. Now we ask him to try again.

* Added tiers to enemy health depending on your coin pouch

The more coins the player has, the harder it gets to kill the monsters. I also made it so instead of trying to use a potion when they're out, it lets them know they're out.